+++ include/soci/firebird/soci-firebird.h
@@ -198,7 +198,7 @@ struct firebird_statement_backend : details::statement_backend
     exec_fetch_result execute(int number) SOCI_OVERRIDE;
     exec_fetch_result fetch(int number) SOCI_OVERRIDE;
 
-    long long get_affected_rows() SOCI_OVERRIDE;
+    int64_t get_affected_rows() SOCI_OVERRIDE;
     int get_number_of_rows() SOCI_OVERRIDE;
     std::string get_parameter_name(int index) const SOCI_OVERRIDE;
 
@@ -231,7 +231,7 @@ protected:
     int rowsFetched_;
     bool endOfRowSet_;
 
-    long long rowsAffectedBulk_; // number of rows affected by the last bulk operation
+    int64_t rowsAffectedBulk_; // number of rows affected by the last bulk operation
 
     virtual void exchangeData(bool gotData, int row);
     virtual void prepareSQLDA(XSQLDA ** sqldap, short size = 10);
@@ -291,7 +291,7 @@ struct firebird_blob_backend : details::blob_backend
 protected:
 
     virtual void open();
-    virtual long getBLOBInfo();
+    virtual int64_t getBLOBInfo();
     virtual void load();
     virtual void writeBuffer(std::size_t offset, char const * buf,
         std::size_t toWrite);
@@ -301,7 +301,7 @@ protected:
     std::vector<char> data_;
 
     bool loaded_;
-    long max_seg_size_;
+    int64_t max_seg_size_;
 };
 
 struct firebird_session_backend : details::session_backend
@@ -315,7 +315,7 @@ struct firebird_session_backend : details::session_backend
     void rollback() SOCI_OVERRIDE;
 
     bool get_next_sequence_value(session & s,
-        std::string const & sequence, long & value) SOCI_OVERRIDE;
+        std::string const & sequence, int64_t & value) SOCI_OVERRIDE;
 
     std::string get_dummy_from_table() const SOCI_OVERRIDE { return "rdb$database"; }
